# Perform convolution with bias addition combined for a specific algorithm 