--- a/dlib/dnn/layers.h
+++ b/dlib/dnn/layers.h
@@ -233,17 +233,19 @@ namespace dlib
                        padding_y_,
                        padding_x_);
 
-            conv(false, output,
-                 sub.get_output(),
-                 filters(params,0));
-
-            // For some reason, doing this is sometimes slower than two separate calls
-            // conv(false, output,
-            //     sub.get_output(),
-            //     filters(params,0),
-            //     biases(params, filters.size()));
             if (use_bias)
-                tt::add(1,output,1,biases(params,filters.size()));
+            {
+                conv(false, output,
+                     sub.get_output(),
+                     filters(params,0),
+                     biases(params, filters.size()));
+            }
+            else
+            {
+                conv(false, output,
+                     sub.get_output(),
+                     filters(params,0));
+            }
         }
 
         template <typename SUBNET>