@arrufat 2023-07-22
Ah, AlexNet doesn't have Batch Normalization layers, so the convolutions there always have biases. As a result, it's unaffected by `fuse_layers`.