@facug91 2023-07-22
> Ah, AlexNet doesn't have Batch Normalization layers, so the convolutions there always have biases. As a result, it's unaffected by `fuse_layers`.

Good point. Anyway, the execution of the combined layer should also change.
I ran some more tests with other models and again, except for AlexNet, they all got better. The second column represents how much faster it is with this modification:

```
alexnet		-1,40%
sqznet1.0	10,79%
sqznet1.1	16,96%
vggnet11	 2,63%
vggnet13	 6,16%
vggnet16	 6,61%
vggnet19	 5,14%
googlenet	 5,57%
resnet18	 7,37%
resnet34	 5,78%
resnet50	11,42%
resnet101	 8,94%
resnet152	 8,51%
darknet19	 7,18%
darknet53	 5,11%
darknet53csp	11,42%
```