@arrufat 2023-07-23
I can confirm similar results on my machine.

This is the pattern that the `fuse_layers` visitor is looking for:
https://github.com/davisking/dlib/blob/a50bd167c4d5b40374706dcdc4369552d57172b8/dlib/dnn/visitors.h#L406

Calling that visitor on a network that doesn't have that pattern (convolution layer followed by affine layer, which has the scaling and bias of the batch norm) won't do anything at all.

I should've paid more attention to the CuDNN documentation when doing all this stuff. Thank you for pointing it out. TBH, I am still surprised that I made it work at all.