@facug91 2023-07-23
I was looking at that PR yesterday, where you added all of this, and it was huge. This is just a very small and specific use case hahah. I'll probably need your help for another PR later, where I plan to modify more things about this, but I'm not entirely sure how to do it.

Thank you for testing and confirming results :)