@facug91 2023-07-27
I think I was able to solve it, but thanks arrufat. When this PR is accepted (or not hahah), I'll create the other one. I depend on this code, but I managed to fuse convolution with bias and activation function, and the improvement is really nice :)